feat(mcp): make missing-dependency hint distro-aware
When a browser fails to launch because of a missing shared library, the
MCP browser factory suggested `sudo npx playwright install-deps`, which
only works on Debian/Ubuntu. Show that suggestion only there; on other
distributions point users at their own package manager (zypper, dnf,
pacman), matching the install-deps validation message.

Move the Debian/Ubuntu detection (isDebianBasedDistro) from
registry/dependencies.ts into @utils/linuxUtils so both the registry and
the MCP tools can share it. Replace the `any`-typed catch clauses and
casts in browserFactory.ts with precise types along the way.
